# implement block decode in assembly