--- a/decode_amd64.go
+++ b/decode_amd64.go
@@ -1,0 +1,8 @@
+// +build !appengine
+// +build gc
+// +build !noasm
+
+package lz4
+
+//go:noescape
+func decodeBlock(dst, src []byte) int