~2x faster

```
name                old time/op    new time/op     delta
UncompressPg1661-4    1.62ms ± 2%     0.61ms ± 1%    -62.14%  (p=0.000 n=10+9)
UncompressDigits-4    98.3µs ± 0%     57.2µs ± 0%    -41.76%  (p=0.000 n=10+10)
UncompressTwain-4     1.05ms ± 1%     0.39ms ± 0%    -62.61%  (p=0.000 n=10+9)
UncompressRand-4      4.16µs ± 1%     4.18µs ± 0%     +0.45%  (p=0.004 n=9+9)

name                old speed      new speed       delta
UncompressPg1661-4   367MB/s ± 2%    968MB/s ± 1%   +164.12%  (p=0.000 n=10+9)
UncompressDigits-4  1.02GB/s ± 0%   1.75GB/s ± 0%    +71.72%  (p=0.000 n=10+10)
UncompressTwain-4    371MB/s ± 1%    992MB/s ± 0%   +167.47%  (p=0.000 n=10+9)
UncompressRand-4    3.94GB/s ± 1%   3.92GB/s ± 0%     -0.45%  (p=0.006 n=9+9)

name                old alloc/op   new alloc/op    delta
UncompressPg1661-4     8.00B ± 0%      4.00B ± 0%    -50.00%  (p=0.000 n=9+8)
UncompressDigits-4     0.00B           0.00B            ~     (all equal)
UncompressTwain-4     2.40B ±100%     0.40B ±150%       ~     (p=0.077 n=10+10)
UncompressRand-4       0.00B           0.00B            ~     (all equal)

name                old allocs/op  new allocs/op   delta
UncompressPg1661-4      0.00            0.00            ~     (all equal)
UncompressDigits-4      0.00            0.00            ~     (all equal)
UncompressTwain-4       0.00            0.00            ~     (all equal)
UncompressRand-4        0.00            0.00            ~     (all equal)
```