@Zariel 2019-02-14
implemting the shortcut decode from the C implementation brings speeds up to ~1.6GB/s on my macbook.

```
name                old time/op    new time/op    delta
UncompressPg1661-4    1.62ms ± 0%    0.37ms ± 0%   -77.31%  (p=0.000 n=9+8)
UncompressDigits-4    98.3µs ± 0%    51.8µs ± 0%   -47.37%  (p=0.000 n=10+10)
UncompressTwain-4     1.04ms ± 0%    0.24ms ± 0%   -76.87%  (p=0.000 n=10+10)
UncompressRand-4      4.19µs ± 0%    4.17µs ± 1%    -0.53%  (p=0.003 n=10+10)

name                old alloc/op   new alloc/op   delta
UncompressPg1661-4    5.60B ±100%    0.90B ±122%      ~     (p=0.053 n=10+10)
UncompressDigits-4     0.00B          0.00B           ~     (all equal)
UncompressTwain-4     2.80B ±100%    0.70B ±100%   -75.00%  (p=0.026 n=10+10)
UncompressRand-4       0.00B          0.00B           ~     (all equal)

name                old allocs/op  new allocs/op  delta
UncompressPg1661-4      0.00           0.00           ~     (all equal)
UncompressDigits-4      0.00           0.00           ~     (all equal)
UncompressTwain-4       0.00           0.00           ~     (all equal)
UncompressRand-4        0.00           0.00           ~     (all equal)

name                old speed      new speed      delta
UncompressPg1661-4   368MB/s ± 0%  1620MB/s ± 0%  +340.75%  (p=0.000 n=9+8)
UncompressDigits-4  1.02GB/s ± 0%  1.93GB/s ± 0%   +89.99%  (p=0.000 n=10+10)
UncompressTwain-4    372MB/s ± 0%  1608MB/s ± 0%  +332.37%  (p=0.000 n=10+10)
UncompressRand-4    3.91GB/s ± 0%  3.93GB/s ± 1%    +0.53%  (p=0.003 n=10+10)
```