@pierrec 2019-02-15
Very nice. Will try to port it to the Go function as well.